Moved gen1 postblocks to gen1 and fixed tests in trainer with removal of auto Concat
