+++ credit/trainers/trainerERA5_Diffusion.py
@@ -3,7 +3,6 @@ import logging
 from collections import defaultdict
 
 import numpy as np
-import torch
 import torch.distributed as dist
 import torch.fft
 import tqdm
@@ -13,7 +12,7 @@ from credit.scheduler import update_on_batch
 from credit.trainers.utils import cycle, accum_log
 from credit.trainers.base_trainer import BaseTrainer
 from credit.data import concat_and_reshape, reshape_only
-from credit.postblock import GlobalMassFixer, GlobalWaterFixer, GlobalEnergyFixer
+from credit.postblock.gen1 import GlobalMassFixer, GlobalWaterFixer, GlobalEnergyFixer
 import torchmetrics
